Move path information out of AppFactory
