+++ src/go/mainfactory/app_factory.go
@@ -2,9 +2,6 @@ package mainfactory
 
 import (
 	"errors"
-	"fmt"
-	"os"
-	"path/filepath"
 
 	"github.com/kkrull/marmot/coremetarepo"
 	"github.com/kkrull/marmot/corerepository"
@@ -13,42 +10,16 @@ import (
 	repository "github.com/kkrull/marmot/userepository"
 )
 
-func DefaultAppFactory() (*AppFactory, error) {
-	if metaRepoPath, pathErr := defaultMetaRepoPath(); pathErr != nil {
-		return nil, pathErr
-	} else {
-		return newAppFactory().ForLocalMetaRepo(metaRepoPath), nil
-	}
-}
-
-func newAppFactory() *AppFactory {
-	return &AppFactory{}
-}
-
-func defaultMetaRepoPath() (string, error) {
-	if homeDir, homeErr := os.UserHomeDir(); homeErr != nil {
-		return "", fmt.Errorf("failed to locate home directory; %w", homeErr)
-	} else {
-		return filepath.Join(homeDir, "meta"), nil
-	}
-}
-
 // Constructs application commands and queries with configurable services.
 type AppFactory struct {
 	MetaDataAdmin    coremetarepo.MetaDataAdmin
-	metaRepoPath     string
 	RepositorySource corerepository.RepositorySource
 }
 
 // Configure a local, file-based meta repo at the specified path
 func (factory *AppFactory) ForLocalMetaRepo(metaRepoPath string) *AppFactory {
-	factory.metaRepoPath = metaRepoPath
 	factory.RepositorySource = svcfs.NewJsonMetaRepo(metaRepoPath)
 	return factory
-}
-
-func (factory *AppFactory) MetaRepoPath() string {
-	return factory.metaRepoPath
 }
 
 /* Administration */
